firedancer: build validator and genesis binaries as libraries
We want to link these into our Firedancer binary for zero configuration
and easy deployment. The main() functions are renamed main1() and a
library is built for each.

Firedancer already links OpenSSL but it's the QUIC variant, and there's
symbol clashes with Solana which also links OpenSSL. We just strip out
the uses of it from Solana since they aren't needed for now anyway.
