+++ Cargo.toml
@@ -100,6 +100,7 @@ members = [
     "udp-client",
     "upload-perf",
     "validator",
+    "validator-fd",
     "version",
     "watchtower",
     "zk-keygen",
@@ -268,7 +269,7 @@ rcgen = "0.10.0"
 reed-solomon-erasure = "6.0.0"
 regex = "1.8.3"
 rolling-file = "0.2.0"
-reqwest = { version = "0.11.17", default-features = false }
+reqwest = { version = "0.11.17", default-features = false, features = ["rustls-tls"] }
 rpassword = "7.2"
 rustc_version = "0.4"
 rustls = { version = "0.20.8", default-features = false }
@@ -354,6 +355,7 @@ solana-thin-client = { path = "thin-client", version = "=1.16.15" }
 solana-tpu-client = { path = "tpu-client", version = "=1.16.15", default-features = false }
 solana-transaction-status = { path = "transaction-status", version = "=1.16.15" }
 solana-udp-client = { path = "udp-client", version = "=1.16.15" }
+solana-validator = { path = "validator", version = "=1.16.15" }
 solana-version = { path = "version", version = "=1.16.15" }
 solana-vote-program = { path = "programs/vote", version = "=1.16.15" }
 solana-zk-keygen = { path = "zk-keygen", version = "=1.16.15" }
