+++ storage-bigtable/Cargo.toml
@@ -17,16 +17,16 @@ bzip2 = { workspace = true }
 enum-iterator = { workspace = true }
 flate2 = { workspace = true }
 futures = { workspace = true }
-goauth = { workspace = true }
+# goauth = { workspace = true }
 http = { workspace = true }
 hyper = { workspace = true }
-hyper-proxy = { workspace = true }
+# hyper-proxy = { workspace = true }
 log = { workspace = true }
 prost = { workspace = true }
 prost-types = { workspace = true }
 serde = { workspace = true }
 serde_derive = { workspace = true }
-smpl_jwt = { workspace = true }
+# smpl_jwt = { workspace = true }
 solana-metrics = { workspace = true }
 solana-sdk = { workspace = true }
 solana-storage-proto = { workspace = true }
@@ -39,12 +39,12 @@ zstd = { workspace = true }
 # openssl is a dependency of the goauth and smpl_jwt crates, but explicitly
 # declare it here as well to activate the "vendored" feature that builds OpenSSL
 # statically...
-[target."cfg(not(windows))".dependencies]
-openssl = { workspace = true, features = ["vendored"] }
+# [target."cfg(not(windows))".dependencies]
+# openssl = { workspace = true, features = ["vendored"] }
 # ...except on Windows to avoid having to deal with getting CI past a build-time
 # Perl dependency
-[target."cfg(windows)".dependencies]
-openssl = { workspace = true, features = [] }
+# [target."cfg(windows)".dependencies]
+# openssl = { workspace = true, features = [] }
 
 [lib]
 crate-type = ["lib"]
